backend/requirements.txt
# PneumoScan Backend Requirements
# Python 3.11+

# Web Framework
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
pydantic==2.5.3

# Machine Learning
tensorflow==2.18.0
numpy==1.26.2
scikit-learn==1.3.2

# Image Processing
Pillow==10.1.0

# Environment Variables
python-dotenv==1.0.0

# Utilities
requests==2.31.0

# Optional: Development & Testing
# pytest==7.4.3
# pytest-cov==4.1.0
# pytest-asyncio==0.21.1
# httpx==0.25.2
# black==23.12.0
# flake8==6.1.0
# mypy==1.7.1
# isort==5.13.2
